backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
SQLAlchemy==2.0.23
alembic==1.12.1
google-generativeai==0.3.2
firebase-admin==6.2.0
httpx==0.25.2
python-magic==0.4.27
Pillow==10.1.0
pandas==2.1.3
numpy==1.25.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
datagovindia==1.1.0
requests==2.31.0